replace icons components into Icons folder
