+++ src/features/Icons/Rules.js
