+++ src/features/Icons/Settings.js
@@ -1,0 +1,9 @@
+export default function Settings({theme}) {
+    const stroke = theme === "theme__light" ? "black" : "white";
+    return (
+        <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M6.26545 3.28808L8.32413 4.66053C8.95167 5.07889 9.79882 4.67845 9.87387 3.92798L10.0766 1.9005C10.1277 1.38929 10.5579 1 11.0717 1H13.0135C13.4902 1 13.9006 1.33646 13.9941 1.80388L14.4388 4.0274C14.5831 4.74885 15.4308 5.07273 16.0194 4.63129L17.7228 3.35371C18.121 3.05512 18.678 3.09471 19.0299 3.4466L20.5534 4.97006C20.9053 5.32196 20.9449 5.87905 20.6463 6.27717L19.3687 7.98061C18.9273 8.56919 19.2511 9.4169 19.9726 9.56119L22.1961 10.0059C22.6635 10.0994 23 10.5098 23 10.9865V13.0135C23 13.4902 22.6635 13.9006 22.1961 13.9941L19.8353 14.4663C19.1497 14.6034 18.8129 15.3839 19.1834 15.9768L20.3246 17.8027C20.5536 18.169 20.5214 18.641 20.2448 18.9729L18.9919 20.4764C18.6648 20.869 18.0942 20.9517 17.669 20.6682L15.6759 19.3395C15.0483 18.9211 14.2012 19.3216 14.1261 20.072L13.9234 22.0995C13.8723 22.6107 13.4421 23 12.9283 23H10.9865C10.5098 23 10.0994 22.6635 10.0059 22.1961L9.56119 19.9726C9.4169 19.2511 8.56919 18.9273 7.98061 19.3687L6.27717 20.6463C5.87905 20.9449 5.32195 20.9053 4.97006 20.5534L3.4466 19.0299C3.09471 18.678 3.05512 18.121 3.35371 17.7228L4.63129 16.0194C5.07273 15.4308 4.74885 14.5831 4.0274 14.4388L1.80388 13.9941C1.33646 13.9006 1 13.4902 1 13.0135V10.9865C1 10.5098 1.33646 10.0994 1.80388 10.0059L4.0274 9.56119C4.74885 9.4169 5.07273 8.56919 4.63129 7.98061L3.35371 6.27717C3.05512 5.87905 3.09471 5.32195 3.4466 4.97006L5.00364 3.41302C5.34071 3.07596 5.86883 3.02366 6.26545 3.28808Z" stroke={stroke} strokeWidth="2"/>
+            <circle cx="12" cy="12" r="1.83333" fill={stroke}/>
+        </svg>
+    )
+}
