+++ src/features/keyboard/Keyboard.js
@@ -1,9 +1,7 @@
 import {Key} from "./key/Key";
 import classes from "./Keyboard.module.scss";
 import Button from "../Button/Button";
-import backspaceLight from '../../assets/light-theme/backspace-light.svg';
-import backspaceDark from '../../assets/dark-theme/backspace-dark.svg';
-import Backspace from "./Icons/Backspace";
+import Backspace from "../Icons/Backspace";
 
 export default function Keyboard({getKeyState, KeyClickHandler, keys, theme, keyDownHandler}) {
     const row1 = keys.filter((el, idx) => idx < 9);
@@ -29,7 +27,7 @@ export default function Keyboard({getKeyState, KeyClickHandler, keys, theme, key
                 {row3.map((key, idx) => <KeyWrap key={idx} idx={idx} children={key}/>)}
                 <div className={classes.btn}>
                     <Button handler={(e) => keyDownHandler({key: 'Backspace'})}>
-                        <img width={34.11} height={22} src={theme === "theme__light" ? backspaceLight : backspaceDark} alt="backspace"/>
+                        <Backspace theme={theme} />
                     </Button>
                 </div>
             </div>
